Move helper method to a shared ByteBuddyUtils class
