+++ sdks/java/core/src/main/java/org/apache/beam/sdk/schemas/utils/POJOUtils.java
@@ -17,7 +17,7 @@
  */
 package org.apache.beam.sdk.schemas.utils;
 
-import static org.apache.beam.sdk.schemas.utils.ByteBuddyUtils.getClassLoadingStrategy;
+import static org.apache.beam.sdk.util.ByteBuddyUtils.getClassLoadingStrategy;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
